pyproject.toml
[tool.poetry]
name = "duplicate-search-engine"
version = "0.1.0"
description = ""
authors = ["raisultan <ki.xbozz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.8.4"
aiohttp-cors = "0.7.0"
python-dotenv = "1.0.0"
pymilvus = "2.3.0"
openai = "0.28.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
